added rjackr's fft code to the vscode version
Added real FFT code and expanded the partition sizes for both app partitions so that the larger version could fit, but also so that OTA could work again.
Remember to build all new elements and erase the chip before uploading it all back.
Also set the default display mode to scroll and to show the IP when the NVS is empty, so users can find the IP if their router doesn't work with mDNS.