Made the boot up more descriptive
Made the boot up more descriptive on what is being setup to better diagnose any startup issues without the need to console in. (I took the mini test clock on vacation to work on it and forgot about the effect from a lack of a known AP for a good 30 seconds).
Now you can see what it's doing as it boots.